tools/pdf_to_markdown/pyproject.toml
[project]
name = "pdf-to-markdown"
version = "0.1.0"
description = "Convertis un fichier pdf composé de scans (images) en fichier markdown"
readme = "README.md"
authors = [
    { name = "Cyril Mammar", email = "mammar.cyril@epitech.eu" }
]
requires-python = ">=3.12"
dependencies = [
    "google-adk==1.1.1",
    "pdf2image==1.17.0",
    "requests==2.31.0",
    "common"
]

[project.scripts]
pdf-to-markdown = "pdf_to_markdown:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
common = { workspace = true }